refactor(ui): split logs ledger and schema ownership
